requirements.txt
Django==2.1.3
django-cors-headers==2.4.0
django-extensions==2.1.4
django-filter==2.1.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
PyJWT==1.6.4
pytz==2018.7
six==1.11.0
